refactor: move class FormulaParser
